+++ account/infrastructure/entry-points/web/src/main/java/ec/kgalarza/bank/dto/AccountInDTO.java
@@ -2,12 +2,14 @@ package ec.kgalarza.bank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AccountInDTO {
 
     private String accountNumber;
     private String accountType;
-    private double onlineBalance;
+    private BigDecimal onlineBalance;
     private boolean status;
     private Long customerId;
