+++ account/infrastructure/entry-points/web/src/main/java/ec/kgalarza/bank/dto/ReportOutDTO.java
@@ -3,6 +3,7 @@ package ec.kgalarza.bank.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,9 +14,9 @@ public class ReportOutDTO {
     private String customerName;
     private String accountNumber;
     private String accountType;
-    private double initialBalance;
-    private double transactionAmount;
-    private double availableBalance;
+    private BigDecimal initialBalance;
+    private BigDecimal transactionAmount;
+    private BigDecimal availableBalance;
     private String transactionDescription;
     private Boolean accountStatus;
