+++ account/infrastructure/entry-points/web/src/main/java/ec/kgalarza/bank/dto/TransactionInDTO.java
@@ -2,12 +2,14 @@ package ec.kgalarza.bank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TransactionInDTO {
 
-    private double initialBalance;
-    private double transactionAmount;
-    private double availableBalance;
+    private BigDecimal initialBalance;
+    private BigDecimal transactionAmount;
+    private BigDecimal availableBalance;
     private String transactionDescription;
     private Long accountId;
 }
