+++ account/infrastructure/entry-points/web/src/main/java/ec/kgalarza/bank/dto/TransactionOutDTO.java
@@ -2,6 +2,7 @@ package ec.kgalarza.bank.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,9 +10,9 @@ public class TransactionOutDTO {
 
     private Long id;
     private LocalDateTime transactionDate;
-    private double initialBalance;
-    private double transactionAmount;
-    private double availableBalance;
+    private BigDecimal initialBalance;
+    private BigDecimal transactionAmount;
+    private BigDecimal availableBalance;
     private String transactionDescription;
 
 }
